Move test modules for manage.py commands into their own directory
