+++ tardis/tardis_portal/tests/management/__init__.py
@@ -1,0 +1,4 @@
+"""
+This package contains tests relating to management commands,
+i.e. command-line tools invoked with manage.py
+"""
