+++ tardis/tardis_portal/tests/management/test_collectstatic.py
